% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geSiteTable.R
\name{geSiteTable}
\alias{geSiteTable}
\title{Build DSRA Object}
\usage{
geSiteTable(
  Hs,
  Water = 0,
  USCS,
  Group = NULL,
  h = 0.5,
  DrID = NULL,
  Vref = 760,
  UniformDistribution = TRUE,
  POP = 100,
  IgnoreModelIntervals = TRUE,
  getSiteLayers = FALSE
)
}
\arguments{
\item{Hs}{Scalar Layer Thickness  in m}

\item{Water}{Scalar. Water Table Depth in \% Hs.  zw = Hs-Water*Hs}

\item{USCS}{Vector. Unified Soil Classification System codes}

\item{Group}{Soil Groups c=("Gravels","Fines","Sands")}

\item{h}{Scalar. Layer Thickness in m}

\item{DrID}{Relative Density  c=("Very Dense","Dense","Compact","Loose","Very Loose"). RD=NULL assumes that...}

\item{Vref}{Reference Velocity in m/s. Default is 760 m/s.}

\item{UniformDistribution}{Boolean. If TRUE, the distribution of the soil properties is assumed to be uniform. If FALSE, the distribution of the soil properties is assumed to be normal.}

\item{POP}{Pre-consolidation pressure in kPa??. OCR= (pm+POP)/pm}

\item{IgnoreModelIntervals}{Boolean. If TRUE, the model intervals are ignored and the soil profile is built using the layer thicknesses provided in Hs.}

\item{getSiteLayers}{Boolean}
}
\value{
DSRA object
}
\description{
Build DSRA Object
}
