% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteProperties.R
\name{getSiteProperties}
\alias{getSiteProperties}
\title{Title}
\usage{
getSiteProperties(
  Hs,
  USCS,
  POP = 100,
  Water = 0,
  NR = 1,
  h = 1,
  levels = c(0.05, 0.5, "mean", 0.95)
)
}
\arguments{
\item{Hs}{Double.}

\item{USCS}{Character.}

\item{POP}{Integer.}

\item{Water}{Double.}

\item{NR}{Integer}

\item{h}{Double}

\item{levels}{list}
}
\value{
List
}
\description{
Title
}
